requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
appnope==0.1.4
asteroid-filterbanks==0.4.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==24.3.0
audioread==3.0.1
braceexpand==0.1.7
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
cytoolz==1.0.1
dataclasses-json==0.6.7
datasets==3.2.0
debugpy==1.8.11
decorator==4.4.2
dill==0.3.8
distro==1.9.0
docopt==0.6.2
editdistance==0.8.1
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.9.0.post1
ffmpeg==1.4
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.9.0
grpcio==1.69.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
hydra-core==1.3.2
HyperPyYAML==1.2.2
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
intervaltree==3.1.0
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jiwer==3.0.5
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
julius==0.2.7
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-huggingface==0.1.2
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
langsmith==0.2.10
lazy_loader==0.4
lhotse==1.29.0
librosa==0.10.2.post1
lightning==2.5.0.post0
lightning-utilities==0.11.9
lilcom==1.8.0
llvmlite==0.43.0
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.24.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nemo-toolkit @ git+https://github.com/NVIDIA/NeMo.git@d98b7cd1b1a9c06d094a8b349bcc2e7b5a7e16a9
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
omegaconf==2.3.0
onnx==1.17.0
openai==1.59.4
optuna==4.1.0
orjson==3.10.13
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
primePy==1.3
proglog==0.1.10
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==3.20.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==18.1.0
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.5.0.post0
pytorch-metric-learning==2.8.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.11.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.1
scikit-learn==1.6.0
scipy==1.15.0
semver==3.0.2
sentence-transformers==3.3.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.0
soxr==0.5.0.post1
speechbrain==1.0.2
SQLAlchemy==2.0.36
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
toolz==1.0.0
torch==2.5.1
torch-audiomentations==0.11.1
torch_pitch_shift==1.2.5
torchaudio==2.5.1
torchmetrics==1.6.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
webdataset==0.2.100
Werkzeug==3.1.3
wget==3.2
wrapt==1.17.0
xxhash==3.5.0
yarl==1.18.3
